Bug 1009390 - [Settings]rename send_feedback and choose_feedback panel to feedback_send and feedback_choose panel
- fix jsodc
