+++ apps/settings/elements/improve_browser_os_choose_feedback.html
@@ -35,6 +35,6 @@
         </li>
       </ul>
     </div>
-    <panel data-path="panels/choose_feedback/panel"></panel>
+    <panel data-path="panels/feedback_choose/panel"></panel>
   </template>
 </element>
