+++ apps/settings/elements/improve_browser_os_send_feedback.html
@@ -58,6 +58,6 @@
         <button id="feedback-done-btn" class="full" data-l10n-id="done">Done</button>
       </menu>
     </form>
-    <panel data-path="panels/send_feedback/panel"></panel>
+    <panel data-path="panels/feedback_send/panel"></panel>
   </template>
 </element>
