+++ apps/settings/jsdoc.json
@@ -7,7 +7,10 @@
             "!apps/settings/js/settings.js",
             "!apps/settings/js/wifi.js",
             "!apps/settings/js/modules/mvvm/list_view.js",
-            "!apps/settings/js/panels/send_feedback/send_feedback.js"
+            "!apps/settings/js/panels/feedback_send/feedback_send.js",
+            "!apps/settings/js/panels/keyboard/core.js",
+            "!apps/settings/js/panels/keyboard_add_layouts/core.js",
+            "!apps/settings/js/panels/keyboard_enabled_layouts/core.js"
     ],
     "options": {
       "destination": "docs/settings"
